Renamer navn på fil - skal ikke bruke ordet ny da det er uklart når og hva som et nytt
