+++ src/models/steg/dinsituasjon/situasjonTyper.ts
